requirements.txt
# Discord
discord.py==1.3.2
PyNaCl==1.3.0
# Audio
ffmpeg-python==0.2.0
# UI
PySide2==5.14.1
qtmodern==0.2.0
